backend/requirements.txt
beautifulsoup4==4.11.1
requests==2.27.1
selenium==4.5.0
pandas==1.5.0
Django==4.1.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
celery==5.2.7
redis==4.3.4
kombu==5.2.4
botocore==1.29.16
mysqlclient==2.1.1